Add FeedbackPRE, fix clipping for photodiode
Added interval between stim display conclusion and reward issue called TASK.FeedbackPRE. Successive rewards have fixed 50ms interval.

mksensor analogy data (e.g., photodiode) sent over realtimedb to mkturk are now stored in /mkturkfiles/daq/ as int16 and not int8 to avoid clipping (arduino data are natively int10 not int8)

+++ public/mksensor/mksensor_usb.js
@@ -496,7 +496,7 @@ function saveTrialData(abuff){
 async function saveAllData(abuff){
   alldata.t0.push(abuff.t0);
   alldata.t.push(abuff.t);
-  alldata.a0.push(new Uint8Array(abuff.ph))
+  alldata.a0.push(new Uint16Array(abuff.ph))
   alldata.nsamples = alldata.nsamples + abuff.ph.length
 
   let blob = new Blob([JSON.stringify(alldata)], { type: 'application/json' });
